large commit, but it adds nothing except stripping unused properties and reorganizing files
