+++ lib/index.tsx
@@ -4,10 +4,8 @@ export * from "./@components";
 export * from "./types";
 export * from "./utils";
 export * from "./hooks";
-export {
-    ComponentRegistry,
-    useChat,
-    useLocale as useLang,
-    useConfigData,
-} from "./providers";
 export { WidgetRoot } from "./Root";
+export { useConfigData } from "./providers/ConfigDataProvider"
+export { useChat } from "./providers/ChatProvider"
+export { useLocale } from "./providers/LocalesProvider";
+export { useWidgetSoundEffects } from "./providers/use-widget-sfx"
