+++ lib/locales/fr.locale.ts
@@ -1,4 +1,4 @@
-import { TranslatableMessages } from "@lib/types";
+import { TranslatableMessages } from ".";
 
 export const frLocale: TranslatableMessages = {
   ok: "D'accord",
