+++ lib/locales/index.ts
@@ -1,1 +1,4 @@
+import { TranslationKeysType } from "./en.locale";
+
 export * from "./helper";
+export type TranslatableMessages = Record<TranslationKeysType, string>;
