--- lib/providers/index.tsx
@@ -1,4 +1,0 @@
-export * from "./ChatProvider";
-export * from "./componentRegistry";
-export * from "./LocalesProvider";
-export { useConfigData } from "./ConfigDataProvider"
