+++ lib/types/index.ts
@@ -1,5 +1,4 @@
 export * from "./helpers";
 export * from "./messages";
 export * from "./options";
-export * from "./translations";
 export * from "./components";
