--- lib/types/translations.ts
@@ -1,3 +1,0 @@
-import { TranslationKeysType } from "@lib/locales/en.locale";
-
-export type TranslatableMessages = Record<TranslationKeysType, string>;
