+++ src/components/tooltip.tsx
@@ -1,8 +1,8 @@
 "use client";
 
-import { cn } from "@lib/utils/cn";
 import * as TooltipPrimitive from "@radix-ui/react-tooltip";
 import * as React from "react";
+import { cn } from "src/utils";
 
 const TooltipProvider = TooltipPrimitive.Provider;
